tests/fixtures/rules/VCR-003/requirements.txt
autogen
semantic-kernel
azure-identity
